code/requirements.txt
# Updated Requirements for Nexora
# Consolidated and verified dependencies based on code review and existing project files.

# Core Framework & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
grpcio==1.59.0
grpcio-tools==1.59.0

# Data Processing & ML/AI
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.0
joblib==1.3.2
faker==20.0.3
# synthgauge removed - not essential, using faker instead
lifelines==0.27.8
apache-beam[gcp]==2.51.0
openlineage-python==1.2.0

# Healthcare Specific
fhir.resources==7.1.0

# Visualization & UI
streamlit==1.28.2
altair==5.1.2
plotly==5.18.0

# Utilities
pytest==7.4.3
pyyaml==6.0.1
# sqlite3 is part of Python standard library
